Adicionando tratamento de erro aos botões
